more fix

+++ .scripts/setup-lf.bash
@@ -16,25 +16,29 @@ done
 
 
 # Use case here for maximum flexibility if we were to change later
-case "$RELEASE_BUILD" in
-    'dev') 
-        git clone https://github.com/lf-lang/lingua-franca.git --branch master --depth 1
-        cd lingua-franca
-        git submodule update --init --recursive
-        ./gradlew buildAll
-        cd .. 
-    ;;
-    *) 
-        python3 ./.scripts/get-lf-executable $RELEASE_BUILD
-        mkdir lingua-franca
-        # While what we have here is tar.gz, lf release bot appear to have a bug and did not gunzip it.
-        # Therefore `tar -xzf` will fail but `tar -xf` will work.
-        # Here, we ignore the actual build name (the original name of the file and the original first directory). 
-        tar -xf lf.tar.gz -C lingua-franca --strip-components 1
-        rm lf.tar.gz
-    ;;
-esac
+if [ ! -d "lingua-franca" ]; then
+    case "$RELEASE_BUILD" in
+        'dev') 
+            git clone https://github.com/lf-lang/lingua-franca.git --branch master --depth 1
+            cd lingua-franca
+            git submodule update --init --recursive
+            ./gradlew buildAll
+            cd .. 
+        ;;
+        *) 
+            python3 ./.scripts/get-lf-executable $RELEASE_BUILD
+            mkdir lingua-franca
+            # While what we have here is tar.gz, lf release bot appear to have a bug and did not gunzip it.
+            # Therefore `tar -xzf` will fail but `tar -xf` will work.
+            # Here, we ignore the actual build name (the original name of the file and the original first directory). 
+            tar -xf lf.tar.gz -C lingua-franca --strip-components 1
+            rm lf.tar.gz
+        ;;
+    esac
+else
+    echo "Directory lingua-franca already exist. Ignoring fetch."
+fi
 
-if [ $EXAMPLE -eq 1 ] ; then
+if [ $EXAMPLE -eq 1 ] && [ ! -d "examples" ] ; then
     git clone https://github.com/lf-lang/examples-lingua-franca.git examples --branch main
 fi
